Correctly shrink selections when reversed
When shrinking selections to the first character, the cursor was being
left on the char *after* the selection if the selection was reversed.
This was causing issues when repeating linewise visual (un)indents,
since one line lower would be included in the indented region each time
it was repeated.

fixes #67